fix: Code refactoring: - Move converters/teds.py into evaluators/teds.py - Move all functions from converters/utils.py into benchmarks/utils.py. - Rename create_xxx_converter() functions. - Rename BenchMarkColumns.DOCLING_VERSION as BenchMarkColumns.CONVERTER_VERSION
