+++ docling_eval/benchmarks/cvat_annotation/preannotate.py
@@ -23,8 +23,7 @@ from docling_eval.benchmarks.cvat_annotation.utils import (
     TableComponentLabel,
     rgb_to_hex,
 )
-from docling_eval.benchmarks.utils import get_binhash
-from docling_eval.converters.utils import insert_images
+from docling_eval.benchmarks.utils import get_binhash, insert_images
 
 # Configure logging
 logging.basicConfig(
