+++ docling_eval/benchmarks/funsd/create.py
@@ -31,15 +31,15 @@ from docling_eval.benchmarks.constants import (
     ConverterTypes,
     EvaluationModality,
 )
-from docling_eval.benchmarks.utils import write_datasets_info
-from docling_eval.converters.conversion import create_image_converter
-from docling_eval.converters.utils import (
+from docling_eval.benchmarks.utils import (
     crop_bounding_box,
     docling_version,
     extract_images,
     from_pil_to_base64uri,
     save_shard_to_disk,
+    write_datasets_info,
 )
+from docling_eval.converters.conversion import create_image_docling_converter
 
 SHARD_SIZE = 1000
 
@@ -447,7 +447,7 @@ def create_funsd_dataset(
     splits: List[str] = ["train", "test"],
     max_items: int = -1,
 ):
-    doc_converter = create_image_converter(do_ocr=True, ocr_lang=["en"])
+    doc_converter = create_image_docling_converter(do_ocr=True, ocr_lang=["en"])
 
     num_train_rows = 0
     num_test_rows = 0
@@ -516,7 +516,7 @@ def create_funsd_dataset(
 
             record = {
                 BenchMarkColumns.CONVERTER_TYPE: ConverterTypes.DOCLING,
-                BenchMarkColumns.DOCLING_VERSION: docling_version(),
+                BenchMarkColumns.CONVERTER_VERSION: docling_version(),
                 BenchMarkColumns.DOC_ID: img_path.stem,
                 BenchMarkColumns.GROUNDTRUTH: json.dumps(true_doc.export_to_dict()),
                 BenchMarkColumns.GROUNDTRUTH_PAGE_IMAGES: true_page_images,
