+++ docling_eval/benchmarks/tableformer_huggingface_otsl/create.py
@@ -22,16 +22,16 @@ from docling_eval.benchmarks.constants import (
     ConverterTypes,
     EvaluationModality,
 )
-from docling_eval.benchmarks.utils import convert_html_table_into_docling_tabledata
-from docling_eval.converters.models.tableformer.tf_model_prediction import (
-    PageTokens,
-    TableFormerUpdater,
-)
-from docling_eval.converters.utils import (
+from docling_eval.benchmarks.utils import (
+    convert_html_table_into_docling_tabledata,
     docling_version,
     extract_images,
     from_pil_to_base64uri,
     save_shard_to_disk,
+)
+from docling_eval.converters.models.tableformer.tf_model_prediction import (
+    PageTokens,
+    TableFormerUpdater,
 )
 from docling_eval.visualisation.visualisations import save_comparison_html
 
@@ -226,7 +226,7 @@ def create_huggingface_otsl_tableformer_dataset(
 
             record = {
                 BenchMarkColumns.CONVERTER_TYPE: ConverterTypes.DOCLING,
-                BenchMarkColumns.DOCLING_VERSION: docling_version(),
+                BenchMarkColumns.CONVERTER_VERSION: docling_version(),
                 BenchMarkColumns.STATUS: str(ConversionStatus.SUCCESS.value),
                 BenchMarkColumns.DOC_ID: str(os.path.basename(filename)),
                 BenchMarkColumns.GROUNDTRUTH: json.dumps(true_doc.export_to_dict()),
@@ -243,7 +243,7 @@ def create_huggingface_otsl_tableformer_dataset(
         else:
             record = {
                 BenchMarkColumns.CONVERTER_TYPE: ConverterTypes.DOCLING,
-                BenchMarkColumns.DOCLING_VERSION: docling_version(),
+                BenchMarkColumns.CONVERTER_VERSION: docling_version(),
                 BenchMarkColumns.STATUS: str(ConversionStatus.FAILURE.value),
                 BenchMarkColumns.DOC_ID: str(os.path.basename(filename)),
                 BenchMarkColumns.GROUNDTRUTH: json.dumps(true_doc.export_to_dict()),
