+++ docling_eval/evaluators/table_evaluator.py
@@ -12,8 +12,8 @@ from pydantic import BaseModel
 from tqdm import tqdm  # type: ignore
 
 from docling_eval.benchmarks.constants import BenchMarkColumns
-from docling_eval.converters.teds import TEDScorer
 from docling_eval.evaluators.stats import DatasetStatistics, compute_stats
+from docling_eval.evaluators.teds import TEDScorer
 
 _log = logging.getLogger(__name__)
